[ADD] 팝업 공지사항 구현
