+++ src/main/resources/templates/layout.html
@@ -369,22 +369,10 @@
             </footer>
         </div>
     </div>
-
-    <!-- 팝업 컨테이너 -->
-    <div id="popup-container">
-        <p id="popup-content"></p>
-        <div class="button-container">
-            <button id="popup-close" class="btn btn-sm btn-light" style="font-size: 10px">닫기</button>
-            <button id="popup-dont-show" class="btn btn-sm btn-secondary" style="font-size: 10px">모든 공지사항 하루 보지 않기</button>
-        </div>
-    </div>
-
     <!--토스트 컨테이너-->
     <div id="toastContainer" style="position: fixed; bottom: 10px; right: 10px; z-index: 9999;"></div>
-
     <script th:src="@{/js/layout.js}"></script>
     <script th:src="@{/js/notification.js}"></script>
-    <script th:src="@{/js/homeNotice.js}"></script>
     <script src="https://cdnjs.cloudflare.com/ajax/libs/Chart.js/2.8.0/Chart.min.js" crossorigin="anonymous"></script>
     </body>
     </html>
